meetmatch_backend/requirements.txt
Django>=6.0
psycopg2-binary
dj-database-url
python-dotenv
django[gis]
djangorestframework
django-cors-headers
requests